[Windows SEH] goto/return/conti/break are abnormal-exits in _try
Per Windows SEH Spec. except _leave or fall-through, all other exits
out of a _try are considered abnormal exits.  in that cases, the first
paremeter passes to its _finally funclet should be TRUE indicating an
abnormal-termination.
Since _leave/fall-through is always Indexed 0 in NormalCleanupDest,
this fix just passes NormalCleanupDest ID as 1st Arg to _finally

Second fix is that in llvm::calculateSEHStateForBlocks(), the incoming State
for "cleanup" could be incorrect when it's passed from abnormal-exits (cases stated above).
the fix is just reset it with Invoke's state.