[Windows SEH] -EHa State propagation for SEME region and several     fixes
The central idea of State propagation along CFG is based on the
following:
  A _try scope is always a SEME (Single Entry Multiple Exits) region
    as jumping into a _try is not allowed
  The single entry must start with a seh_try_begin() invoke with a
    correct State number that is the initial state of the SEME.
  Through control-flow, state number is propagated into all blocks.
  Side exits marked by seh_try_end() will unwind to parent state via
    existing SEHUnwindMap[].
  Side exits can ONLY jump into parent scopes (lower state number).
  Thus, when a block succeeds various states from its predecessors,
    the lowest State triumphs others.
  If some exits flow to unreachable, propagation on those paths terminate,
    not affecting remaining blocks.

For CPP code, object lifetime region is usually a SEME as SEH _try.
However there is one rare exception: jumping into a lifetime that has
Dtor but has no Ctor is warned but allowed:i

   note: jump bypasses variable with a non-trivial destructor

In this case, this region is actually a MEME (multiple entry multiple
exits).  Our solution is to inject a seh_try_begin in that side entry
block to ensure a correct State.

This changes also fix the following:
(1) Under -EHa, suppress push/popTerminate() (from noexcept/noTHrow) so
    that HW exceptions can be passed through.
(2) Intercept present normal-cleanup block with try_/scope_end() to mark
    side-exits of SEME region
(3) Volatilizing block: also include calls to MemIntrinsic().
(4) Block State reporting: need to report blocks with State -1. Or it
    could be subsumed into adjacent States.
(5) CalculateCXXStateNumbers() Bug: the CatchPad state in
    EHPadStateMap[]  should be CatchLow, not TryLow.